Update table component for pagination and navigation
Updated the 'ReactTableProps' and 'Table' components to populate and navigate movie data in pages. Added 'moviesInfo' and 'setData' and a flag 'showNavigation' in 'ReactTableProps'. Updated 'Table' to extract pagination info from URLs and added logic to load data for different pages. Fetch URL has also been fixed. Styling for the extra elements has been kept intact. This provides the user with a better understanding of the amount of data and easier navigation.